+++ src/om_bootstrap/modal.cljs
@@ -1,104 +1,58 @@
 (ns om-bootstrap.modal
   "IN PROGRESS work on a modal component. Depends on a fade mixin."
-  (:require [cljs.core.async :as a :refer [chan put! close!]]
-            [om.core :as om]
-            [om-bootstrap.mixins :refer [set-listener-mixin]]
+  (:require [om.core :as om]
             [om-bootstrap.types :as t]
             [om-tools.core :refer-macros [defcomponentk]]
             [om-tools.dom :as d :include-macros true]
-            [om-tools.mixin :refer-macros [defmixin]]
-            [schema.core :as s])
-  (:require-macros [cljs.core.async.macros :refer [go-loop]]
-                   [schema.macros :as sm]))
+            [schema.core :as s]
+            [om-bootstrap.util :as u])
+  (:require-macros [schema.macros :as sm]))
 
 ;; ## Schema
-;;
-;; TODO: Should we make on-request-hide a channel that the user
-;; optionally supplies?
-;;
-;; TODO: Figure out how the hell the d/div, etc are defined. I don't
-;; like this pattern of having to explicitly include a "children"
-;; field. We should always take the properties as the first argument,
-;; and then optionally take any number of children after that.... but
-;; does that work with om/build? I don't think so. This is one to ask
-;; David Nolen about.
 
 (def Modal
   "Options for the modal."
-  {:title (s/named s/Any "Either a valid react component OR a string.")
-   :backdrop? s/Bool
-   :keyboard? s/Bool
-   :close-button? s/Bool
-   :animate? s/Bool
-   :children (s/named s/Any "Child elements, if any.")
-   :on-request-hide (sm/=> s/Any s/Any)})
+  {:header s/Any
+   :footer s/Any
+   (s/optional-key :keyboard?) s/Bool
+   (s/optional-key :close-button?) s/Bool
+   (s/optional-key :visible?) s/Bool
+   (s/optional-key :animate?) s/Bool})
 
-(sm/defn render-header :- t/Component
-  "Renders the header for the modal."
-  [{:keys [close-button? on-request-hide title]} :- Modal]
-  (let [close-button (when close-button?
-                       (d/button {:type "button"
-                                  :class "close"
-                                  :aria-hidden true
-                                  :on-click on-request-hide}
-                                 "&times;"))]
-    (d/div {:class "modal-header"}
-           close-button
-           (if (.isValidComponent js/React title)
-             title
-             (d/h4 {:class "modal-title"} title)))))
-
-(sm/defn render-backdrop :- t/Component
-  "Renders the backdrop behind the modal."
-  [{:keys [animation? backdrop? on-request-hide] :- Modal} modal-elem]
-  (let [klasses {:modal-backdrop true
-                 :fade animation?
-                 :in (or (not animation?)
-                         (not (.querySelectorAll js/document)))}]
-    (d/div {:class (d/class-set klasses)
-            :on-click (when backdrop?
-                        (fn [e]
-                          (when (= (.-target e)
-                                   (.-currentTarget e))
-                            (on-request-hide))))
-            :ref "backdrop"}
-           modal-elem)))
-
-;; TODO: Look up the bootstrap modal mixin here!
-;;
-;; https://gist.github.com/insin/8449696
-#_
-(defcomponentk modal
-  "Component to render a Bootstrap modal to the DOM."
-  [[:data animation? on-request-hide {keyboard? true} :as modal] :- Modal
-   owner state]
-  (:mixins fade-mixin set-listener-mixin)
-  (init-state [_] {:keyup (chan)})
-  (will-mount [_]
-              (let [c (:keyup @state)]
-                (go-loop []
-                  (let [item (a/<! c)]
-                    (if (and keyboard? (.-keycode item))
-                      (on-request-hide)))
-                  (recur))))
-  (did-mount [_] (.set-listener owner "keyup" (:keyup @state)))
+(defcomponentk modal*
+  "Component that renders a Modal. Manages it's own toggle state"
+  [owner state]
+  (init-state [_]
+    {:visible? (get-in (om/get-props owner) [:opts :visible?])})
+  (will-receive-props [this next-props]
+    (let [last-props (om/get-props owner)
+          last-visible? (om/get-state owner [:visible?])
+          next-visible? (get-in next-props [:opts :visible?])]
+      (when (not= last-visible? next-visible?)
+        (om/set-state! owner [:visible?] next-visible?))))
   (render [_]
-          (d/div {} (d/p {} "Placeholder."))
-          #_(let [style {:display "block"}
-                  klasses {:modal true
-                           :fade animation?
-                           :in (or (not animation?)
-                                   (not (.querySelectorAll js/document)))}
-                  modal-elem (d/div {:title nil
-                                     :tab-index "-1"
-                                     :role "dialog"
-                                     :class (d/class-set klasses)
-                                     :ref "modal"}
-                                    (d/div {:class "modal-dialog"}
-                                           (d/div {:class "modal-content"}
-                                                  (when (:title modal)
-                                                    (render-header modal))
-                                                  (:children modal))))]
-              (if (:backdrop? modal)
-                (render-backdrop modal modal-elem)
-                modal-elem))))
+    (let [{:keys [opts children]} (om/get-props owner)
+          [bs props] (t/separate Modal opts {:bs-class "modal"})
+          classes {:modal true
+                   :fade true
+                   :in (om/get-state owner [:visible?])}]
+      (d/div (u/merge-props props
+                            {:class (d/class-set classes)})
+        (d/div {:class "modal-dialog"}
+          (d/div {:class "modal-content"}
+            (d/div {:class "modal-header"}
+              (when (:close-button? bs)
+                (d/button {:type         "button"
+                           :class        "close"
+                           :aria-hidden  true
+                           :on-click (fn [_] (om/set-state! owner [:visible?] false))}
+                          "×"))
+              (u/clone-with-props (:header bs) {:class "modal-title"}))
+            (d/div {:class "modal-body"}
+              children)
+            (d/div {:class "modal-footer"}
+              (:footer bs))))))))
+
+(sm/defn modal
+  [opts :- Modal & children]
+  (->modal* {:opts opts :children children}))