+++ CHANGELOG.md
@@ -1,5 +1,10 @@
-## 0.3.5 (1/27/2015)
+## 0.3.6 (2/5/2015)
 
+- Actually build modal component (https://github.com/racehub/om-bootstrap/pull/53)
+
+## 0.3.5 (2/2/2015)
+
+- Modal component, thanks to @dignati! (https://github.com/racehub/om-bootstrap/pull/50)
 - Support for collapsible panels, thanks to @dignati (https://github.com/racehub/om-bootstrap/pull/45)
 
 ## 0.3.4 (1/27/2015)